render_service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
pillow==10.4.0
requests==2.32.3
